crates/sel4/bitfield-types: Move Bitfield type into sel4-sys
