+++ crates/sel4/sys/src/bf/mod.rs
@@ -1,6 +1,8 @@
 use core::fmt;
 
-use sel4_bitfield_types::Bitfield;
+pub(crate) mod types;
+
+use types::Bitfield;
 
 include!(concat!(env!("OUT_DIR"), "/types.rs"));
 include!(concat!(env!("OUT_DIR"), "/shared_types.rs"));
